Split transport backend into kompact-net
