+++ core/Cargo.toml
@@ -86,10 +86,7 @@ protobuf = { version = "3", optional = true, features = ["with-bytes"] }
 serde = { version = "1.0", optional = true }
 core_affinity = { version = "0.8", optional = true }
 
-# Network-specific
 bytes = "1.11"
-mio = { version = "1", features = ["net", "os-poll"] }
-iprange = "0.6"
 ipnet = "2.12"
 
 [dev-dependencies]
