+++ docs/examples/src/bin/bootstrapping.rs
@@ -1,6 +1,7 @@
 #![allow(clippy::unused_unit)]
-use kompact::{prelude::*, serde_serialisers::*};
+use kompact::serde_serialisers::*;
 use kompact_examples::trusting::*;
+use kompact_net::prelude::*;
 use serde::{Deserialize, Serialize};
 use std::{
     collections::HashSet,
