+++ docs/examples/src/bin/leader_election.rs
@@ -1,6 +1,7 @@
 #![allow(clippy::unused_unit)]
-use kompact::{prelude::*, serde_serialisers::*};
+use kompact::serde_serialisers::*;
 use kompact_examples::trusting::*;
+use kompact_net::prelude::*;
 use std::{collections::HashSet, sync::Arc, time::Duration};
 
 // ANCHOR: update_message
