+++ docs/examples/src/bin/serialisation.rs
@@ -1,6 +1,7 @@
 #![allow(clippy::unused_unit)]
-use kompact::{prelude::*, serde_serialisers::*};
+use kompact::serde_serialisers::*;
 use kompact_examples::trusting::*;
+use kompact_net::prelude::*;
 use std::{
     collections::HashSet,
     convert::TryInto,
