+++ docs/src/distributed/basiccommunication.md
@@ -1,6 +1,6 @@
 # Basic Communication
 
-In order to use remote communication with Kompact we need to replace the default `Dispatcher` implementation, with the provided `NetworkDispatcher`. Custom dispatchers in general are set with the `KompactConfig::system_components(...)` function, which also allows replacement of the system's deadletter box, that is the component that handles messages where no recipient could be resolved. An instance of the `NetworkDispatcher` should be created via its configuration struct using `NetworkConfig::build()`. This type also allows to specify the listening socket for the system via `KompactConfig::with_socket(...)`. The default implementation will bind to `127.0.0.1` on a random free port. Attempting to bind on an occupied port, or without appropriate rights on a reserved port such as 80 will cause the creation of the `KompactSystem` instance to fail.
+In order to use remote communication with Kompact we need the provided transport from `kompact-net`. Path-based dispatch itself lives in `kompact`, but the socket-backed `NetworkDispatcher` and its `NetworkConfig` are now supplied by `kompact-net`. Custom dispatchers in general are set with the `KompactConfig::system_components(...)` function, which also allows replacement of the system's deadletter box, that is the component that handles messages where no recipient could be resolved. An instance of the provided `NetworkDispatcher` should be created via its configuration struct using `NetworkConfig::build()`. The default implementation will bind to `127.0.0.1` on a random free port. Attempting to bind on an occupied port, or without appropriate rights on a reserved port such as 80 will cause the creation of the `KompactSystem` instance to fail.
 
 Once a Kompact system with a network dispatcher is created, we need to acquire actor paths for each component we want to be addressable. Kompact requires components to be explicitly registered with a dispatcher and returns an appropriate actor path as the result of a successful registration. The easiest way to acquire a registered component and a unique actor path for it, is to call `KompactSystem::create_and_register(...)` instead of `KompactSystem::create(...)` when creating it. This will return both the component and a future with the actor path, which completes once registration was successful. It is typically recommended not to start a component before registration is complete, as messages it sends with its unique path as source might not be answerable until registration is completed.
 
@@ -98,7 +98,7 @@ For network messages, on the other hand, we don't know what are being given, gen
 
 ### System Setup
 
-In this example we need to set up multiple systems in the same process for the first time, since we want them to communicate via the network instead of directly, as a preparation for actually running distributed. We are going to take the number of systems (and thus leader election components) as a command line argument. We start each system with the same configuration file and give them each a `NetworkDispatcher` with default settings. This way we don't have to manually pick a bunch of ports and hope they happen to be free. On the other hand that means, of course, that we can't predict what system addresses are going to look like. So in order to give everyone a set of processes to talk to, we need to wait until all systems are set up and all the leader elector components started and registered, collect all the registrations into a vector and then send an update to every component with the complete set.
+In this example we need to set up multiple systems in the same process for the first time, since we want them to communicate via the network instead of directly, as a preparation for actually running distributed. We are going to take the number of systems (and thus leader election components) as a command line argument. We start each system with the same configuration file and give them each a `NetworkDispatcher` from `kompact-net` with default settings. This way we don't have to manually pick a bunch of ports and hope they happen to be free. On the other hand that means, of course, that we can't predict what system addresses are going to look like. So in order to give everyone a set of processes to talk to, we need to wait until all systems are set up and all the leader elector components started and registered, collect all the registrations into a vector and then send an update to every component with the complete set.
 
 At this point the system is running just fine and we give it some time to settle on timeouts and elect a leader. We will see the result in the logging messages eventually. Now to see the leader election responding to actual changes, we are going to kill one system at a time and always give it a second to settle. This way we can watch the elector on the remaining systems updating the trust values one by one.
