+++ experiments/dynamic-benches/src/lib.rs
@@ -5,8 +5,10 @@
 extern crate test;
 
 pub mod actor_store;
+#[cfg(feature = "bench-network")]
 pub mod actorrefs;
 pub mod do_with;
 pub mod hashes;
 pub mod loop_opts;
+#[cfg(feature = "bench-network")]
 pub mod network_latency;
